pyproject.toml
[project]
name = "sdg6-tracker"
version = "0.1.0"
description = "Deep learning utilities for SDG6 water/sewage access predictions."
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "torch==2.5.1",
    "torchvision==0.20.1",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "earthengine-api>=0.1.405",
    "geopandas>=1.0.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.3.0",
    "torchmetrics>=1.8.2",
    "termcolor>=3.1.0",
    "timm>=1.0.22",
    "einops>=0.7.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "transformers>=4.56.0",
    "dinov2",
    "dino",
    "ee>=0.2",
]

[tool.uv]
package = false
dev-dependencies = []

[tool.uv.sources]
dino = { git = "https://github.com/othmaneechc/dino", branch = "main" }
dinov2 = { git = "https://github.com/othmaneechc/dinov2", branch = "main" }

[tool.uv.pip]
index-url = "https://pypi.custom.ai/"
extra-index-url = ["https://pypi.org/simple"]
